10th commit - Added __str__ to Query model class - Refactoring exceptions handeling in the views

+++ Project/query_api/views.py
@@ -1,19 +1,16 @@
-import json
-import time, requests
-from random import randint,choice
-from django.db import transaction
 from django.shortcuts import get_object_or_404
 from rest_framework import viewsets, status
 from rest_framework.response import Response
-from rest_framework.exceptions import APIException
+from rest_framework.exceptions import APIException, ValidationError
 
 from query_api.models import Query
 from query_api.serializers import QuerySerializer
 from query_api.tasks import send_query
 
+
 class QueryViewSet(viewsets.ViewSet):
     """
-    A views set for posting data to /query endpoint
+    A viewset for the /query endpoint
     """
 
     def create(self, request):
@@ -23,23 +20,35 @@ class QueryViewSet(viewsets.ViewSet):
             serializer = QuerySerializer(data=data)
             serializer.is_valid(raise_exception=True)
             query = serializer.save()
-            query_id = query.id
 
-            #Celery task
-            #send_query.delay(query_id)
+            #Celery task that gets the "result" and saves it in the DB
+            send_query.delay(query.id)
 
             return Response(
-                {"query_id":query_id}, 
-                status=status.HTTP_201_CREATED
-            )
+                {"query_id": query.id}, 
+                status=status.HTTP_201_CREATED)
 
+        except ValidationError as e:
+            return Response({"error": e.detail}, status=status.HTTP_400_BAD_REQUEST)
+        
         except APIException as e:
-            query.delete()
-            return Response({"error":str(e.detail)})
+            if query.id:
+                query.delete()
+            return Response(
+                {"error": e.detail},status=e.status_code)
+        
+        except Exception as e:
+            print(e)
+            if query.id:
+                query.delete()
+            return Response(
+                {"error": "An unexpected error occurred."},
+                status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
 
 class ResultViewSet(viewsets.ViewSet):
+    """A viewset for the /result endpoint"""
 
     def retrieve(self, request, pk):
         try:
@@ -49,8 +58,15 @@ class ResultViewSet(viewsets.ViewSet):
             return Response(serializer.data,status=status.HTTP_200_OK)
         
         except APIException as e:
+            return Response(
+                {"error": e.detail},status=e.status_code)
+        
+        except Exception as e:
             print(e)
-            return Response({"error":"An unexpected error occured"},status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            return Response(
+                {"error": "An unexpected error occurred."},
+                status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+    
     
     def partial_update(self, request, pk):
         try:
@@ -62,16 +78,26 @@ class ResultViewSet(viewsets.ViewSet):
             serializer.save()
 
             return Response(serializer.data,status=status.HTTP_200_OK)
+        
         except APIException as e:
+            return Response(
+                {"error": e.detail},status=e.status_code)
+        
+        except Exception as e:
             print(e)
-            return Response({"error":"An unexpected error occured"},status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
+            return Response(
+                {"error": "An unexpected error occurred."},
+                status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+    
 class HistoryViewSet(viewsets.ViewSet):
+    """A viewset for the /history endpoint"""
 
     def list(self,request):
-        cadastral_number = request.query_params['cadastral_number']
+        #Enable filtering by cadastral-number
+        cadastral_number = request.query_params.get('cadastral_number')
         queryset = Query.objects.all()
 
+        #Filter the base queryset
         if cadastral_number:
             queryset = queryset.filter(cadastral_number=cadastral_number)
             query = get_object_or_404(queryset)
@@ -87,8 +113,13 @@ class HistoryViewSet(viewsets.ViewSet):
             return Response({"error": "No queries found"}, status=status.HTTP_404_NOT_FOUND)
 
         except APIException as e:
+            return Response(
+                {"error": e.detail},status=e.status_code)
+        
+        except Exception as e:
             print(e)
-            return Response({"error":"An unexpected error occured"},status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-
+            return Response(
+                {"error": "An unexpected error occurred."},
+                status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
